+++ simtools/applications/compare_cumulative_psf.py
@@ -86,8 +86,8 @@ import numpy as np
 import yaml
 
 import simtools.utils.general as gen
-from simtools import io_handler
 from simtools.configuration import configurator
+from simtools.io import io_handler
 from simtools.model.telescope_model import TelescopeModel
 from simtools.ray_tracing import RayTracing
 from simtools.visualization import visualize
