+++ simtools/applications/generate_corsika_histograms.py
@@ -170,10 +170,10 @@ from pathlib import Path
 import numpy as np
 
 import simtools.utils.general as gen
-from simtools import io_handler
 from simtools.configuration import configurator
 from simtools.corsika import corsika_histograms_visualize
 from simtools.corsika.corsika_histograms import CorsikaHistograms
+from simtools.io import io_handler
 
 logger = logging.getLogger()
