+++ simtools/applications/make_regular_arrays.py
@@ -40,8 +40,9 @@ from pathlib import Path
 import astropy.units as u
 
 import simtools.utils.general as gen
-from simtools import db_handler, io_handler
+from simtools import db_handler
 from simtools.configuration import configurator
+from simtools.io import io_handler
 from simtools.layout.layout_array import LayoutArray
