+++ simtools/applications/sim_showers_for_trigger_rates.py
@@ -68,8 +68,8 @@ from pathlib import Path
 
 import astropy.units as u
 
-from simtools import io_handler
 from simtools.configuration import configurator
+from simtools.io import io_handler
 from simtools.simulator import Simulator
 from simtools.utils import general as gen
