+++ simtools/applications/validate_camera_efficiency.py
@@ -57,9 +57,9 @@ import logging
 from pathlib import Path
 
 import simtools.utils.general as gen
-from simtools import io_handler
 from simtools.camera_efficiency import CameraEfficiency
 from simtools.configuration import configurator
+from simtools.io import io_handler
 from simtools.model.telescope_model import TelescopeModel
