+++ simtools/corsika/corsika_config.py
@@ -7,7 +7,7 @@ import numpy as np
 from astropy.io.misc import yaml
 
 import simtools.utils.general as gen
-from simtools import io_handler
+from simtools.io import io_handler
 from simtools.layout.layout_array import LayoutArray
 from simtools.utils import names
 from simtools.utils.general import collect_data_from_yaml_or_dict
