+++ simtools/corsika/corsika_histograms.py
@@ -13,11 +13,12 @@ from corsikaio.subblocks import event_header, get_units_from_fields, run_header
 from ctapipe.io import write_table
 from eventio import IACTFile
 
-from simtools import io_handler, version
+from simtools import version
+from simtools.io import io_handler
+from simtools.io.io_hdf5 import fill_hdf5_table
 from simtools.utils.general import (
     collect_data_from_yaml_or_dict,
     convert_2D_to_radial_distr,
-    fill_hdf5_table,
     rotate,
     save_dict_to_file,
 )
