+++ simtools/corsika/corsika_runner.py
@@ -3,11 +3,11 @@ import os
 from copy import copy
 from pathlib import Path
 
-from simtools import io_handler
 from simtools.corsika.corsika_config import (
     CorsikaConfig,
     MissingRequiredInputInCorsikaConfigData,
 )
+from simtools.io import io_handler
 from simtools.utils import names
 from simtools.utils.general import collect_data_from_yaml_or_dict
