+++ simtools/model/array_model.py
@@ -1,7 +1,8 @@
 import logging
 from copy import copy
 
-from simtools import db_handler, io_handler
+from simtools import db_handler
+from simtools.io import io_handler
 from simtools.layout.layout_array import LayoutArray
 from simtools.model.telescope_model import TelescopeModel
 from simtools.simtel.simtel_config_writer import SimtelConfigWriter
